chatbot_backend/requirements.txt
flask
requests
python-dotenv
langchain-google-genai
google-generativeai
flask_cors